requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
python-telegram-bot==21.12
